+++ CHANGELOG.md
@@ -1,14 +1,18 @@
 ## 0.0.1
+
 -Initial Package Realease (Beta)
 
 ## 0.0.2
+
 -Package Correction (Beta)
 
 ## 0.0.3
+
 -Updated Kuda Service with Response (Beta)
 -Added Model to Services
 
 ## 0.0.4
+
 -Updated Kuda Service with Response (Beta)
 -Added More Model to Services
 -Added Transactions History Log
@@ -17,6 +21,7 @@
 -Updated Service Type
 
 ## 0.0.5
+
 -Updated Service typed
 -Added Kuda Savings API
 -Updated kuda_bank.dart
@@ -24,6 +29,7 @@
 -Updated Models
 
 ## 0.0.6
+
 -Updated Service typed
 -Added Kuda Savings API
 -Updated kuda_bank.dart
@@ -34,6 +40,7 @@
 -added kuda savings
 
 ## 0.0.7
+
 -Updated Service Type
 -Added Kuda_Card.dart
 -Updated Model
@@ -44,14 +51,25 @@
 -Updated all libraries
 
 ## 0.0.8
+
 -Updated README.md
 -Updated all library files
 -Code Fixes
 -Updated API Calls
 
 ## 0.0.9
+
 -Updated Example with 2 new details
 -Updated Model Library
 -Updated API library
 -Updated Services Library
 -Bug fixes and improment
+
+## 0.1.0
+
+-Updated service type library
+-update kuda_bank services
+-updated example program
+-updated all library
+-chnaged model to response
+-updated UI for example.